edit vagas page
style.css: put variables for primary and secondary colors; override color on some bootstrap classes; create efects for hover on icons and links in nav and footer. create effect of background bleed for the botton of the page.
vagas.html: added link to fontawesome icons on <header> and added icons for contrast, user and logout links. added footer.